requirements.txt
discord.py>=2.5.0
PyMySQL>=1.1.1
requests>=2.32.3
google-genai>=1.39.0
aiohttp>=3.7.4.post0
dataclass_wizard>=0.22.2
pyotp>=2.6.0
cc-ampapi>=1.3.0